+++ src/gestions_de_stock/LigneCommande.java
@@ -9,6 +9,7 @@ import Table_data.Stock_view;
 import Table_data.ligneCommande_view;
 import Table_data.navigation_button;
 
+import javax.swing.*;
 import javax.swing.event.ListSelectionEvent;
 import javax.swing.event.ListSelectionListener;
 import javax.swing.plaf.basic.BasicInternalFrameUI;
@@ -440,12 +441,15 @@ public class LigneCommande extends javax.swing.JInternalFrame {
             jLabel9.setText("Veuiller remplir Tous les Champs !!!");
             jLabel9.setVisible(true);
         }else {
+            int option = JOptionPane.showConfirmDialog(null, "Êtes-vous sûr de vouloir supprimer cette Ligne de Commande?", "Confirmation", JOptionPane.YES_NO_OPTION);
 
-            ligneCommande_view.DeleteLigneCmd(jTextField23,jTable1);
-            jLabel10.setText("Ligne Supprimer !!!");
-            jLabel10.setVisible(true);
-            jLabel9.setVisible(false);
-            ligneCommande_view.clear(jTextField23,jTextField26, jComboBox1, jComboBox2, jTable1);
+            if (option == JOptionPane.YES_OPTION) {
+                ligneCommande_view.DeleteLigneCmd(jTextField23, jTable1);
+                jLabel10.setText("Ligne Supprimer !!!");
+                jLabel10.setVisible(true);
+                jLabel9.setVisible(false);
+                ligneCommande_view.clear(jTextField23, jTextField26, jComboBox1, jComboBox2, jTable1);
+            }
         }
     }//GEN-LAST:event_jButton3ActionPerformed
