+++ src/gestions_de_stock/lignelivraison.java
@@ -8,6 +8,7 @@ import Table_data.Impression_Exportation;
 import Table_data.lignelivraison_view;
 import Table_data.navigation_button;
 
+import javax.swing.*;
 import javax.swing.event.ListSelectionEvent;
 import javax.swing.event.ListSelectionListener;
 import javax.swing.plaf.basic.BasicInternalFrameUI;
@@ -438,12 +439,15 @@ public class lignelivraison extends javax.swing.JInternalFrame {
             jLabel9.setText("Veuiller remplir Tous les Champs !!!");
             jLabel9.setVisible(true);
         }else {
+            int option = JOptionPane.showConfirmDialog(null, "Êtes-vous sûr de vouloir supprimer cette ligne de livraison?", "Confirmation", JOptionPane.YES_NO_OPTION);
 
-            lignelivraison_view.DeleteLigneliv(jTextField23,jTable1);
-            jLabel10.setText("Ligne Supprimer !!!");
-            jLabel10.setVisible(true);
-            jLabel9.setVisible(false);
-            lignelivraison_view.clear(jTextField23,jTextField26, jComboBox1, jComboBox2, jTable1);
+            if (option == JOptionPane.YES_OPTION) {
+                lignelivraison_view.DeleteLigneliv(jTextField23, jTable1);
+                jLabel10.setText("Ligne Supprimer !!!");
+                jLabel10.setVisible(true);
+                jLabel9.setVisible(false);
+                lignelivraison_view.clear(jTextField23, jTextField26, jComboBox1, jComboBox2, jTable1);
+            }
         }
     }//GEN-LAST:event_jButton3ActionPerformed
